Refactored config parsing and flag parsing
