+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/PingCAP-QE/metrics-checker/cmd"
+
+func main() {
+	cmd.Execute()
+}
